A function that prints a triangle using #s

+++ 0x04-more_functions_nested_loops/100-prime_factor.c
@@ -8,9 +8,9 @@
  */
 int main(void)
 {
-	long long int num;
-	long long int maxNum;
-	long long int i;
+	long int num;
+	long int maxNum;
+	long int i;
 
 	num = 612852475143;
 	maxNum = -1;
